+++ app.js
@@ -71,167 +71,11 @@ mailer.extend(app, {
 require('./routes/index')(app, passport);
 require('./routes/product')(app, passport);
 require('./routes/user')(app, passport);
+require('./routes/admin')(app, passport);
+require('./routes/order')(app, passport);
 
 
-// Set up route for salesman
-app.get('/salesman/order', function(req, res) {
-    Order.find({}).exec(function(err, orders) {
-        if (err) {
-            console.log("Error:", err);
-        } else {
-            res.render('order', { order_items: orders, title: 'RexShop', salesman: true });
-        }
-    });
-});
 
-app.get('/salesman/order/add', function(req, res) {
-    res.render('add_order', { title: 'RexShop', salesman: true });
-});
-app.get('/salesman/order/:id', function(req, res) {
-    Order.findById(req.params.id).exec(function(err, order) {
-        if (err) {
-            console.log("Error:", err);
-        } else {
-            res.render('show_order_detail', { order_item: order, title: 'RexShop', salesman: true });
-        }
-    });
-});
-
-app.get('/salesman/order/edit/:id', function(req, res) {
-    Order.findOne({ _id: req.params.id }).exec(function(err, order) {
-        if (err) {
-            return next(err);
-        } else {
-            res.render('edit_order', { order_item: order, title: 'RexShop', salesman: true })
-        }
-    });
-});
-
-app.post('/salesman/order/delete/:id', function(req, res) {
-    Order.findByIdAndRemove({ _id: req.params.id }).exec(function(err, order) {
-        if (err) {
-            return next(err);
-        } else {
-            console.log("Deleted");
-            res.redirect("/salesman/order");
-        }
-    });
-});
-
-app.post('/salesman/order/update/:id', function(req, res) {
-    Order.findByIdAndUpdate({ _id: req.params.id }, {
-        $set: {
-            customer_id: req.body.customer_id,
-            cost: req.body.cost,
-            date: req.body.date,
-            status: req.body.select_picker,
-            count: req.body.count,
-            item_list: req.body.item_list
-
-        }
-    }, { new: true }, function(err, order) {
-        if (err) {
-            console.log(err);
-        }
-
-        res.redirect("/salesman/order/" + order._id);
-    });
-});
-
-app.post('/salesman/order/save', function(req, res) {
-    var customer_id = req.body.customer_id;
-    var date = req.body.date;
-    var status = req.body.select_picker;
-    var item = req.body.item;
-    var count = req.body.count;
-    var count2 = Number(count);
-    Product.findById({ _id: item }).exec(function(err, product) {
-        if (err) {
-            console.log(err);
-        }
-
-        var order = new Order({
-            customer_id: customer_id,
-            date: date,
-            status: status,
-            cost: '',
-            count: count2,
-            item_list: []
-        });
-        order.cost = product.market_price * count2;
-        order.item_list.push({ item: item, amount: count2 });
-        order.save(function(err) {
-            if (err) {
-                console.log(err);
-                res.render('add_order', { title: 'RexShop', salesman: true });
-            } else {
-                console.log("Successfully created an order.");
-                res.redirect('/salesman/order/' + order._id);
-            }
-        });
-
-    });
-});
-
-
-//Set up route for admin
-
-app.get('/admin', function(req, res) {
-    res.render('index_admin', { admin: true, title: 'RexShop' });
-});
-
-app.get('/admin/manage_account', function(req, res) {
-    User.find({}).exec(function(err, users) {
-        if (err) {
-            console.log("Error:", err);
-        } else {
-            res.render('account_table', { items: users, title: 'RexShop', admin: true });
-        }
-    });
-});
-
-app.get('/admin/manage_account/:id', function(req, res) {
-    User.findById(req.params.id).exec(function(err, user) {
-        if (err) {
-            console.log("Error:", err);
-        } else {
-            res.render('show_account_detail', { item: user, title: 'RexShop', admin: true });
-        }
-    });
-});
-
-app.get('/admin/manage_account/edit/:id', function(req, res) {
-    User.findOne({ _id: req.params.id }).exec(function(err, user) {
-        if (err) {
-            return next(err);
-        } else {
-            res.render('edit_account', { item: user, title: 'RexShop', admin: true })
-        }
-    });
-});
-
-app.post('/admin/manage_account/update/:id', function(req, res) {
-    var name = req.body.name;
-    var address = req.body.address;
-    var phone = req.body.phone;
-    var role = req.body.select_picker;
-    var email = req.body.email;
-    User.findByIdAndUpdate({ _id: req.params.id }, {
-        $set: {
-            name: name,
-            address: address,
-            phone_number: phone,
-            role: role,
-            email: email
-        }
-    }, { new: false }, function(err, user) {
-        if (err) {
-            console.log(err);
-        }
-
-        res.redirect("/admin/manage_account/" + user._id);
-    });
-});
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
@@ -250,4 +94,3 @@ app.use(function(err, req, res, next) {
 });
 
 module.exports = app;
-x
