+++ views/show_account_detail.hbs
@@ -1,3 +1,4 @@
+
 <div class="show_account_detail">
             <div class="container">
                 <div class="row">
@@ -22,7 +23,7 @@
                                     <td>{{item.phone_number}}</td>
                                 </tr>
                                 <tr>
-                                    <td>Role</td>
+                                    <td class="">Role</td>
                                     <td>{{item.role}}</td>
                                 </tr>
                                 <tr>
